Bug 280430 - Version override type Triplet is handled incorrectly

+++ org.eclipse.buckminster.ui/src/java/org/eclipse/buckminster/ui/actions/OpenQueryAction.java
@@ -13,8 +13,10 @@ import java.io.OutputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
 
+import org.eclipse.buckminster.core.helpers.FileUtils;
 import org.eclipse.buckminster.download.DownloadManager;
 import org.eclipse.buckminster.runtime.IOUtils;
+import org.eclipse.buckminster.runtime.URLUtils;
 import org.eclipse.buckminster.ui.Messages;
 import org.eclipse.buckminster.ui.UiPlugin;
 import org.eclipse.buckminster.ui.UiUtils;
@@ -75,8 +77,8 @@ public class OpenQueryAction implements IWorkbenchWindowActionDelegate
 	{
 		IPreferenceStore preferences = UiPlugin.getDefault().getPreferenceStore();
 		Shell shell = m_workbenchWindow.getShell();
-		InputDialog askURL = new InputDialog(shell, null, Messages.url_for_query_with_colon, preferences
-				.getString(LAST_CQUERY_URL), new URLValidator());
+		InputDialog askURL = new InputDialog(shell, null, Messages.url_for_query_with_colon,
+				preferences.getString(LAST_CQUERY_URL), new URLValidator());
 
 		if(askURL.open() != Window.OK)
 			return;
@@ -92,23 +94,29 @@ public class OpenQueryAction implements IWorkbenchWindowActionDelegate
 		preferences.setValue(LAST_CQUERY_URL, urlStr);
 		try
 		{
-			URL url = new URL(urlStr);
-			File tempFile = File.createTempFile(BlankQueryAction.TEMP_FILE_PREFIX, ".cquery"); //$NON-NLS-1$
-			tempFile.deleteOnExit();
+			URL url = URLUtils.normalizeToURL(urlStr);
+			File cqueryFile = FileUtils.getFile(url);
+			if(cqueryFile == null)
+			{
+				File tempFile = File.createTempFile(BlankQueryAction.TEMP_FILE_PREFIX, ".cquery"); //$NON-NLS-1$
+				tempFile.deleteOnExit();
+				OutputStream output = null;
+				try
+				{
+					output = new FileOutputStream(tempFile);
+					DownloadManager.readInto(url, null, output, null);
+				}
+				finally
+				{
+					IOUtils.close(output);
+				}
+				cqueryFile = tempFile;
+			}
+
 			IWorkbench workbench = PlatformUI.getWorkbench();
 			IEditorDescriptor ed = workbench.getEditorRegistry().getDefaultEditor("buckminster.cquery"); //$NON-NLS-1$
-			OutputStream output = null;
-			try
-			{
-				output = new FileOutputStream(tempFile);
-				DownloadManager.readInto(url, null, output, null);
-			}
-			finally
-			{
-				IOUtils.close(output);
-			}
 			m_workbenchWindow.getActivePage().openEditor(
-					new ExternalFileEditorInput(tempFile, new Path(url.toURI().getPath()).lastSegment(), urlStr),
+					new ExternalFileEditorInput(cqueryFile, new Path(url.toURI().getPath()).lastSegment(), urlStr),
 					ed.getId());
 		}
 		catch(Exception e)
